Add functionality to toggle background image/video #14
